Remove GINTERPOLANT initialization.  Change minimal to use catenoid exact soln.
The GINTERPOLANT crap is a hack.  Remove test 9 of fish, which tested it.

In poissonfunctions.c|h, make InitialState() use a DM not a DMDALocalInfo.

Switched DMCreateGlobalVector() to DMGetGlobalVector() in fish.c and minimal.c,
but note there must be a bug in the VI solver because it throws error in
obstacle.c.

Replace -mse_laplace with -mse_power X option, which is a nice parameter.

Remove manufactured solution in minimal.c.  Replace with catenoid solution.
Note -snes_grid_sequence, or other effective initialization strategy is
required for catenoid exact solution.

Remove RHS function f(x,y) in minimal.c.